requirements.txt
# Python dependencies for conflict-tracker (backend/scrapers)
fastapi
uvicorn
pydantic
python-multipart
requests
selenium
webdriver-manager
beautifulsoup4
lxml
psycopg2-binary
python-dotenv
pytest
bcrypt
PyJWT
